backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
python-multipart==0.0.20
sqlalchemy==2.0.40
greenlet
asyncpg==0.30.0
pgvector==0.3.6
pydantic-settings==2.8.1
email-validator==2.2.0
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sentence-transformers==3.4.1
llama-cpp-python==0.3.8
transformers==4.51.3
torch==2.6.0
pypdf==5.4.0
python-docx==1.1.2
httpx==0.28.1
